Replace magic-nix-cache-action (#44)
Why
===

The `DeterminateSystems/magic-nix-cache-action` action has been
[deprecated](https://determinate.systems/posts/magic-nix-cache-free-tier-eol/)
and no longer works.


What changed
============

This PR replaces it with the `nix-community/cache-nix-action` action,
which is actively maintained. It also replaces the
`DeterminateSystems/nix-installer-action` action with the
`nixbuild/nix-quick-install-action` action, which is required for the
cache action to work. The cache has been configured to restore based on
the runner's OS and architecture, so it should work on any runner.

Test plan
=========

Run CI.

---------

Co-authored-by: Devon Stewart <blast@hardchee.se>